backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.1
Django==5.2.6
django-cors-headers==4.9.0
django-environ==0.12.0
django-extensions==4.1
djangorestframework==3.16.1
dnspython==2.8.0
filelock==3.19.1
frozenlist==1.7.0
geoip2==5.1.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
maxminddb==2.8.2
multidict==6.6.4
netaddr==1.3.0
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==5.9.8
psycopg2-binary==2.9.10
pycparser==2.23
python-dateutil==2.9.0.post0
python-nmap==0.7.1
python-whois==0.9.5
pytz==2025.2
redis==6.4.0
requests==2.32.5
requests-file==2.1.0
scapy==2.6.1
shodan==1.31.0
six==1.17.0
sqlparse==0.5.3
tldextract==5.3.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.11.0
xlsxwriter==3.2.9
yarl==1.20.1

# Machine Learning and Security Analytics
scikit-learn==1.6.1
numpy==1.26.4
pandas==2.2.3
tensorflow==2.18.1
joblib==1.5.2

# Advanced analytics and NLP
statsmodels==0.14.0
transformers==4.34.0
spacy==3.8.0
gensim==4.2.0
pyod==1.1.3


# STIX/TAXII Support
stix2==3.0.1
taxii2-client==2.3.0
# mitreattck-python==2.1.0  # commented out: invalid package name / not found on PyPI; verify and re-add if needed

# Vulnerability Management
python-libnmap==0.7.3
nvdlib==0.7.7
cvss==3.0

# SIEM and Log Management
elasticsearch==8.14.0
loguru==0.7.3
pyyaml==6.0.2
